+++ frontend/app/(tabs)/settings.tsx
@@ -1,6 +1,5 @@
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import {
-  Alert,
   Animated,
   Image,
   Platform,
@@ -12,9 +11,7 @@ import {
   useWindowDimensions,
 } from 'react-native';
 import { useMutation, usePaginatedQuery, useQuery } from 'convex/react';
-import { Switch } from 'react-native';
-import { requestPermissionAndSyncToken } from '../../hooks/usePushNotifications';
-import { useRouter, type Href } from 'expo-router';
+import { useRouter } from 'expo-router';
 import { api } from 'convex/_generated/api';
 import { useAuth } from '../../hooks/useAuth';
 import { useEntranceAnimation } from '../../hooks/useEntranceAnimation';
@@ -35,19 +32,17 @@ function yearToOrdinal(gradYear: number): string {
 
 type TabId = 'listings' | 'saved';
 
-const PROFILE_EDIT: Href = '/profile/edit';
+const PROFILE_EDIT = '/profile/edit';
+const ACCOUNT_SETTINGS = '/account-settings';
+
 export default function SettingsScreen() {
   const router = useRouter();
   const { width } = useWindowDimensions();
   const isWeb = Platform.OS === 'web';
-  const { isAuthenticated, isLoading, signOut } = useAuth();
+  const { isAuthenticated, isLoading } = useAuth();
   const entranceStyle = useEntranceAnimation();
-  const signOutInProgressRef = useRef(false);
 
   const [activeTab, setActiveTab] = useState<TabId>('listings');
-  const [isSigningOut, setIsSigningOut] = useState(false);
-  const [messageNotificationsValue, setMessageNotificationsValue] = useState(true);
-  const [isUpdatingMessageNotifications, setIsUpdatingMessageNotifications] = useState(false);
   const profile = useQuery(api.profiles.getCurrentProfile, isAuthenticated && !isWeb ? {} : 'skip');
   const myListings = useQuery(api.listings.getMyListings, isAuthenticated && !isWeb ? {} : 'skip');
   const savedArgs = isAuthenticated && !isWeb && activeTab === 'saved' ? {} : 'skip';
@@ -57,16 +52,6 @@ export default function SettingsScreen() {
     loadMore: loadMoreSavedListings,
   } = usePaginatedQuery(api.savedListings.getMySavedListings, savedArgs, { initialNumItems: 20 });
   const toggleSavedListing = useMutation(api.savedListings.toggleSavedListing);
-  const messageNotificationsEnabled = useQuery(
-    api.users.getMessageNotificationsEnabled,
-    isAuthenticated && !isWeb ? {} : 'skip'
-  );
-  const updateMessageNotificationsEnabled = useMutation(
-    api.users.updateMessageNotificationsEnabled
-  );
-  const recordPushToken = useMutation(api.pushNotifications.recordPushToken);
-  const removePushToken = useMutation(api.pushNotifications.removePushToken);
-  const deleteAccount = useMutation(api.users.deleteAccount);
   const { mappedUrls: avatarUrls } = useResolvedImageUrls(
     profile?.picture ? [profile.picture] : []
   );
@@ -83,145 +68,6 @@ export default function SettingsScreen() {
       router.replace('/auth/login?returnTo=%2Fsettings' as never);
     }
   }, [isAuthenticated, isLoading, isWeb, router]);
-
-  useEffect(() => {
-    if (typeof messageNotificationsEnabled === 'boolean') {
-      setMessageNotificationsValue(messageNotificationsEnabled);
-    }
-  }, [messageNotificationsEnabled]);
-
-  const handleAuthAction = async () => {
-    if (!isAuthenticated) {
-      router.replace('/auth/login?returnTo=%2Fsettings' as never);
-      return;
-    }
-
-    if (signOutInProgressRef.current) {
-      return;
-    }
-
-    try {
-      signOutInProgressRef.current = true;
-      setIsSigningOut(true);
-      await signOut();
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Failed to sign out';
-      Alert.alert('Sign Out Failed', message);
-    } finally {
-      signOutInProgressRef.current = false;
-      setIsSigningOut(false);
-    }
-  };
-
-  const handleDeleteAccount = () => {
-    Alert.alert(
-      'Delete account',
-      'Are you sure? This will permanently delete your account and all your data. You can always sign back up later.',
-      [
-        { text: 'Cancel', style: 'cancel' },
-        {
-          text: 'Delete account',
-          style: 'destructive',
-          onPress: async () => {
-            try {
-              await deleteAccount({});
-            } catch (error) {
-              const message = error instanceof Error ? error.message : 'Failed to delete account';
-              Alert.alert('Delete Account Failed', message);
-              return;
-            }
-
-            try {
-              await signOut();
-            } catch (error) {
-              const details = error instanceof Error ? `\n\nDetails: ${error.message}` : '';
-              Alert.alert(
-                'Account Deleted',
-                `Your account was deleted, but we could not sign you out automatically. Please sign out manually.${details}`
-              );
-            }
-          },
-        },
-      ]
-    );
-  };
-
-  const handleMessageNotificationsToggle = async (value: boolean) => {
-    if (isUpdatingMessageNotifications) return;
-
-    const previousValue = messageNotificationsValue;
-    setMessageNotificationsValue(value);
-    setIsUpdatingMessageNotifications(true);
-
-    try {
-      if (value) {
-        let permissionGranted = false;
-        try {
-          permissionGranted = await requestPermissionAndSyncToken(recordPushToken);
-        } catch (error) {
-          setMessageNotificationsValue(previousValue);
-          const message =
-            error instanceof Error ? error.message : 'Unable to enable notifications right now.';
-          Alert.alert('Notification Update Failed', message);
-          return;
-        }
-
-        if (!permissionGranted) {
-          setMessageNotificationsValue(previousValue);
-          Alert.alert(
-            'Notification Update Failed',
-            'Push permission was not granted. Enable notifications in system settings and try again.'
-          );
-          return;
-        }
-
-        await updateMessageNotificationsEnabled({ enabled: true });
-      } else {
-        let removePushTokenSucceeded = false;
-        let removePushTokenError: unknown = null;
-        try {
-          await removePushToken({});
-          removePushTokenSucceeded = true;
-        } catch (error) {
-          removePushTokenError = error;
-        }
-
-        try {
-          await updateMessageNotificationsEnabled({ enabled: false });
-        } catch (error) {
-          const updatePreferenceMessage =
-            error instanceof Error ? error.message : 'Failed to update notification preference';
-
-          if (removePushTokenSucceeded) {
-            Alert.alert(
-              'Notification partially updated',
-              `This device push token was removed, but we could not save your notification preference.\n\nDetails: ${updatePreferenceMessage}`
-            );
-            return;
-          }
-
-          setMessageNotificationsValue(previousValue);
-          const removeTokenMessage =
-            removePushTokenError instanceof Error
-              ? removePushTokenError.message
-              : 'Failed to remove this device push token.';
-
-          Alert.alert(
-            'Notification Update Failed',
-            `We could not disable notifications.\n\nToken removal: ${removeTokenMessage}\nPreference update: ${updatePreferenceMessage}`
-          );
-          return;
-        }
-      }
-    } catch (error) {
-      setMessageNotificationsValue(previousValue);
-      const message =
-        error instanceof Error ? error.message : 'Failed to update notification preference';
-      Alert.alert('Notification Update Failed', message);
-    } finally {
-      setIsUpdatingMessageNotifications(false);
-    }
-  };
 
   if (isWeb) {
     return (
@@ -266,9 +112,18 @@ export default function SettingsScreen() {
           </Text>
           <Pressable
             style={({ pressed }) => [styles.primaryButton, pressed && styles.buttonPressed]}
-            onPress={() => router.push(PROFILE_EDIT)}
+            onPress={() => router.push(PROFILE_EDIT as never)}
           >
             <Text style={styles.primaryButtonText}>Set up profile</Text>
+          </Pressable>
+          <Pressable
+            style={({ pressed }) => [styles.settingsRow, pressed && styles.buttonPressed]}
+            onPress={() => router.push(ACCOUNT_SETTINGS as never)}
+            accessibilityRole="button"
+            accessibilityLabel="Account settings"
+          >
+            <Text style={styles.settingsRowLabel}>Account settings</Text>
+            <Text style={styles.settingsRowChevron}>›</Text>
           </Pressable>
         </Animated.View>
       </ScrollView>
@@ -317,7 +172,7 @@ export default function SettingsScreen() {
             <Text style={styles.bioText}>{profile.bio}</Text>
             <Pressable
               style={({ pressed }) => [styles.editIcon, pressed && { opacity: 0.7 }]}
-              onPress={() => router.push(PROFILE_EDIT)}
+              onPress={() => router.push(PROFILE_EDIT as never)}
               accessibilityLabel="Edit profile"
               accessibilityRole="button"
             >
@@ -327,7 +182,7 @@ export default function SettingsScreen() {
         ) : (
           <Pressable
             style={({ pressed }) => [styles.addBioRow, pressed && { opacity: 0.7 }]}
-            onPress={() => router.push(PROFILE_EDIT)}
+            onPress={() => router.push(PROFILE_EDIT as never)}
           >
             <Text style={styles.addBioText}>Add a bio...</Text>
             <Text style={styles.editIconText}>✎</Text>
@@ -345,22 +200,18 @@ export default function SettingsScreen() {
         */}
       </Animated.View>
 
-      <View style={styles.notificationsSection}>
-        <Text style={styles.sectionTitle}>Notifications</Text>
-        <View style={styles.notificationRow}>
-          <Text style={styles.notificationLabel}>Message notifications</Text>
-          <Switch
-            value={messageNotificationsValue}
-            onValueChange={(value) => void handleMessageNotificationsToggle(value)}
-            disabled={isUpdatingMessageNotifications || messageNotificationsEnabled === undefined}
-            trackColor={{ false: colors.border, true: colors.primary }}
-            thumbColor={colors.white}
-          />
+      <Pressable
+        style={({ pressed }) => [styles.settingsRowCard, pressed && styles.buttonPressed]}
+        onPress={() => router.push(ACCOUNT_SETTINGS as never)}
+        accessibilityRole="button"
+        accessibilityLabel="Open settings"
+      >
+        <View style={styles.settingsRowTextBlock}>
+          <Text style={styles.settingsRowLabel}>Settings</Text>
+          <Text style={styles.settingsRowSubtext}>Notifications, sign out, and account</Text>
         </View>
-        <Text style={styles.notificationHint}>
-          Get notified when someone messages you about a listing.
-        </Text>
-      </View>
+        <Text style={styles.settingsRowChevron}>›</Text>
+      </Pressable>
 
       <View style={styles.tabs}>
         <Pressable
@@ -444,33 +295,6 @@ export default function SettingsScreen() {
           )}
         </View>
       )}
-
-      <View style={styles.footer}>
-        <Pressable
-          style={({ pressed }) => [
-            styles.secondaryButton,
-            pressed && styles.buttonPressed,
-            (isSigningOut || isLoading) && styles.buttonDisabled,
-          ]}
-          onPress={handleAuthAction}
-          disabled={isSigningOut || isLoading}
-        >
-          <Text style={styles.secondaryButtonText}>
-            {isSigningOut ? 'Signing out...' : 'Sign out'}
-          </Text>
-        </Pressable>
-        <Pressable
-          style={({ pressed }) => [
-            styles.deleteButton,
-            pressed && styles.buttonPressed,
-            isLoading && styles.buttonDisabled,
-          ]}
-          onPress={handleDeleteAccount}
-          disabled={isLoading}
-        >
-          <Text style={styles.deleteButtonText}>Delete account</Text>
-        </Pressable>
-      </View>
     </ScrollView>
   );
 }
@@ -583,30 +407,41 @@ const styles = StyleSheet.create({
     ...typography.footnote,
     color: colors.text,
   },
-  notificationsSection: {
+  settingsRowCard: {
+    flexDirection: 'row',
+    alignItems: 'center',
     backgroundColor: colors.surface,
     borderRadius: borderRadius.xl,
     borderWidth: 1,
     borderColor: colors.border,
-    padding: spacing.xl,
-    gap: spacing.sm,
+    paddingVertical: spacing.md,
+    paddingHorizontal: spacing.lg,
+    gap: spacing.md,
   },
-  sectionTitle: {
-    ...typography.heading,
-    color: colors.textDark,
+  settingsRowTextBlock: {
+    flex: 1,
+    gap: 2,
   },
-  notificationRow: {
+  settingsRow: {
     flexDirection: 'row',
     alignItems: 'center',
     justifyContent: 'space-between',
+    marginTop: spacing.md,
+    paddingVertical: spacing.sm,
   },
-  notificationLabel: {
-    ...typography.body,
-    color: colors.text,
+  settingsRowLabel: {
+    ...typography.subhead,
+    fontWeight: '600',
+    color: colors.textDark,
   },
-  notificationHint: {
+  settingsRowSubtext: {
     ...typography.footnote,
     color: colors.muted,
+  },
+  settingsRowChevron: {
+    fontSize: 22,
+    color: colors.muted,
+    fontWeight: '300',
   },
   bioRow: {
     flexDirection: 'row',
@@ -699,20 +534,6 @@ const styles = StyleSheet.create({
     color: colors.text,
     textAlign: 'center',
   },
-  footer: {
-    marginTop: spacing.lg,
-    gap: spacing.md,
-  },
-  deleteButton: {
-    alignSelf: 'flex-start',
-    paddingVertical: spacing.md,
-    paddingHorizontal: 0,
-  },
-  deleteButtonText: {
-    ...typography.subhead,
-    color: colors.destructive,
-    fontWeight: '600',
-  },
   primaryButton: {
     backgroundColor: colors.primary,
     borderRadius: borderRadius.md,
@@ -745,8 +566,5 @@ const styles = StyleSheet.create({
   buttonPressed: {
     opacity: 0.92,
     transform: [{ scale: 0.99 }],
-  },
-  buttonDisabled: {
-    opacity: 0.7,
   },
 });
